requirements/base.txt
Django==1.5.1
South==0.8.1
bpython==0.12
django-braces==1.2.2
django-model-utils==1.4.0
django_annoying==0.7.7
django_crispy_forms==1.4.0
logutils==0.3.3
requests==2.0.0